Moving the unit test to the correct folder
